wip: moved generation table creation on python side

+++ src/utils/parseCSVs.py
@@ -1,4 +1,3 @@
-import pandas
 import sqlite3
 import sys
 import json
@@ -14,19 +13,7 @@ class FilteredObject:
         for key, value in args.items():
             setattr(self, key, value)
 
-def filter_consumption_per_capita_data(data: Dict[list, Any]):
-    """
-    Filter unuseful consumption data
-    """
-    filtered_data = list() # this is a list
-    for country_data in data:
-        # check if first element exists, check if array is longer then 1, if first element is string, and if the third elemnt is not NaN
-        if len(country_data) >= 3:
-            if isinstance(country_data[0], str) and isinstance(int(country_data[2]), int) and not math.isnan(float(country_data[3])):
-                data_obj = dict(country=country_data[0], year=int(country_data[2]), consumption=float(country_data[3]))
-                filtered_data.append(data_obj)
-    return filtered_data
-
+# This function will be used for both consumption and consumption per capita data
 def filter_consumption_data(data: Dict[list, Any]):
     """
     Filter unuseful consumption data
@@ -40,9 +27,22 @@ def filter_consumption_data(data: Dict[list, Any]):
                 filtered_data.append(data_obj)
     return filtered_data
 
+def filter_generation_data(data: Dict[list, Any]):
+    """
+    Filter unuseful generation data
+    """
+    filtered_data = list() # this is a list
+    for country_data in data:
+        # check if first element exists, check if array is longer then 1, if first element is string, and if the third elemnt is not NaN
+        if len(country_data) >= 3:
+            if isinstance(country_data[0], str) and isinstance(int(country_data[2]), int) and not math.isnan(float(country_data[3])):
+                data_obj = dict(country=country_data[0], year=int(country_data[2]), generation=float(country_data[3]))
+                filtered_data.append(data_obj)
+    return filtered_data
+
 def process_csvs(table_name: TableType, cursor: sqlite3.Cursor) -> Dict[str, Any]:
     """
-    Processes CSV data provided as a JSON string and inserts it into a SQLite3 database.
+    Processes CSV data provided from the name of the table JSON file and inserts it into the SQLite3 database.
 
     :param data: JSON string representing CSV data.
     :param db_path: Path to the SQLite3 database.
@@ -57,7 +57,7 @@ def process_csvs(table_name: TableType, cursor: sqlite3.Cursor) -> Dict[str, Any
                 if json_file_path:
                     with open(json_file_path, 'r', encoding='utf-8') as json_data:
                         data = json.load(json_data)
-                        filtered_data = filter_consumption_per_capita_data(data)
+                        filtered_data = filter_consumption_data(data)
                         cursor.execute('''
                                     CREATE TABLE IF NOT EXISTS ElectricityConsumptionPerCapitaTable (
                                         id INTEGER PRIMARY KEY AUTOINCREMENT,
@@ -73,16 +73,15 @@ def process_csvs(table_name: TableType, cursor: sqlite3.Cursor) -> Dict[str, Any
 
                         # Insert filtered data
                         for row in filtered_data:
-                            try:
-                                # Ensure the row has at least 3 elements to match table columns
-                                if len(row) >= 3:
+                            if len(row) >= 3:
+                                try:
                                     updated_at = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
-                                    cursor.execute(insert_query, (row[0], row[1], row[2], updated_at))
-                                    print('added', row[1])
-                            except sqlite3.IntegrityError as e:
-                                print(f"Error inserting row {row}: {e}")
+                                    cursor.execute(insert_query, (row['country'], row['year'], row['consumption'], updated_at))
+                                except sqlite3.IntegrityError as e:
+                                    print(f"Error inserting row {row}: {e}")
 
                         # Commit the transaction
+                        print("Populated ElectricityConsumptionPerCapitaTable")
                         conn.commit()
                     os.remove(json_file_path)
             case "ElectricityConsumptionTable":
@@ -106,17 +105,48 @@ def process_csvs(table_name: TableType, cursor: sqlite3.Cursor) -> Dict[str, Any
 
                         # Insert filtered data
                         for row in filtered_data:
-                            try:
-                                # Ensure the row has at least 3 elements to match table columns
-                                if len(row) >= 3:
+                            if len(row) >= 3:
+                                try:
                                     updated_at = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
-                                    cursor.execute(insert_query, (row[0], row[1], row[2], updated_at))
-                                    print('added', row[1])
-                            except sqlite3.IntegrityError as e:
-                                print(f"Error inserting row {row}: {e}")
+                                    cursor.execute(insert_query, (row['country'], row['year'], row['consumption'], updated_at))
+                                except sqlite3.IntegrityError as e:
+                                    print(f"Error inserting row {row}: {e}")
 
                         # Commit the transaction
                         conn.commit()
+                        print("Populated ElectricityConsumptionTable")
+                    os.remove(json_file_path)
+            case "ElectricityGenerationTable":
+                json_file_path = os.path.join(current_dir, "temp", "ElectricityGenerationTable.json")
+                if json_file_path:
+                    with open(json_file_path, 'r', encoding='utf-8') as json_data:
+                        data = json.load(json_data)
+                        filtered_data = filter_generation_data(data)
+                        cursor.execute('''
+                                    CREATE TABLE IF NOT EXISTS ElectricityGenerationTable (
+                                        id INTEGER PRIMARY KEY AUTOINCREMENT,
+                                        country VARCHAR(255),
+                                        year INTEGER NOT NULL,
+                                        generation FLOAT,
+                                        createdAt DATETIME DEFAULT CURRENT_TIMESTAMP,
+                                        updatedAt DATETIME DEFAULT CURRENT_TIMESTAMP
+                                    )
+                                    ''')
+
+                        insert_query = 'INSERT INTO ElectricityGenerationTable (country, year, generation, updatedAt) VALUES (?, ?, ?, ?)'
+
+                        # Insert filtered data
+                        for row in filtered_data:
+                            if len(row) >= 3:
+                                try:
+                                    updated_at = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+                                    cursor.execute(insert_query, (row['country'], row['year'], row['generation'], updated_at))
+                                except sqlite3.IntegrityError as e:
+                                    print(f"Error inserting row {row}: {e}")
+
+                        # Commit the transaction
+                        conn.commit()
+                        print("Populated ElectricityGenerationTable")
                     os.remove(json_file_path)
             case _:
                 print(f"An error occured during processing data for {table_name}")
